requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
